Create emu65xx.js
oh my god, keep out of the scope stuff, it is such a waste of time.  Really I want to create a framework which handles that both intelligently and intuitively.  I have looked and looked but haven't found any good existing frameworks that do that.  The couple of times I have rolled my own, I spend so much time dealing with this and never have found a good solution.  Either they end up bad benchmarks, limited enough to not make them intuitive or just plain throw in the garbage.  Javascript is a beautiful monster that ever tries to come of age but can never shake the scope separation between the prototype and the constructor.  That one thing makes it monstrous and ugly.  I feel sorry for Javascript.  If only we could somehow get everyone to adopt NodeJS with DOM as a new standard scripting language supported by all browsers and OS; then life would be great.